Cleanup Traits & Database Imports

+++ src/Database/Migrations/2014_10_12_500000_adapt_users_table_to_login_providers.php
@@ -14,9 +14,11 @@ class AdaptUsersTableToLoginProviders extends Migration
     public function up()
     {
         if (config('login-providers.providers.email')) {
-            Schema::table('users', function (Blueprint $table) {
-                $table->dropColumn('password');
-            });
+            if (Schema::hasColumn('users', 'password')) {
+                Schema::table('users', function (Blueprint $table) {
+                    $table->dropColumn('password');
+                });
+            }
         }
     }
